refactor: :recycle: Solved issue with cards
Deleted itemCard and created a card for each type instead, makes it more maintainable. Also updated the types passed down, cleaned-up folder structure and atom design patterns
