+++ components/molecules/index.ts
@@ -1,0 +1,6 @@
+export * from "./logo";
+export * from "./navigationItems";
+
+export * from "./filmCard";
+export * from "./personCard";
+export * from "./specieCard";
