+++ components/molecules/logo/index.ts
@@ -1,0 +1,1 @@
+export * from "./logo";
