+++ components/molecules/navigationItems/index.ts
@@ -1,0 +1,1 @@
+export * from "./navigationItems";
