+++ components/molecules/personCard/index.ts
@@ -1,0 +1,1 @@
+export * from "./personCard";
