+++ components/molecules/personCard/personCard.tsx
@@ -1,0 +1,52 @@
+import { Favourite } from "@/components/favourite";
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card";
+import { Person } from "@/lib/types";
+
+interface PersonCardProps {
+  person: Person;
+}
+
+/**
+ * Person Card component
+ *
+ * @param {PersonCardProps} { person }
+ * @returns {JSX.Element}
+ */
+export const PersonCard = ({ person }: PersonCardProps): JSX.Element => {
+  const { name, gender, height, hair_color, films, url } = person;
+  console.log("🚀 ~ file: personCard.tsx:23 ~ PersonCard ~ person:", person)
+  return (
+    <Card className="w-[350px]">
+      <CardHeader>
+        <CardTitle className="flex justify-between gap-2">
+          <span>{name}</span>
+          <Favourite url={url} />
+        </CardTitle>
+        <CardDescription>
+          <span>In episode </span>
+          {films.map((item: string, index: number) => {
+            const film = item.split("/");
+            const filmNumber = film.at(film.length - 2);
+            const isLastElement = index === films.length - 1;
+            return (
+              <span key={item}>{`${filmNumber}${
+                isLastElement ? "" : ", "
+              }`}</span>
+            );
+          })}
+        </CardDescription>
+      </CardHeader>
+      <CardContent>
+        <p>{`Gender: ${gender}`}</p>
+        <p>{`Height: ${height}`}</p>
+        <p>{`Hair color: ${hair_color}`}</p>
+      </CardContent>
+    </Card>
+  );
+};
