+++ components/molecules/specieCard/index.ts
@@ -1,0 +1,1 @@
+export * from "./specieCard";
